refactor(operator): thick hooks + form/dialog split for product detail (#279)
Split product-detail-page.tsx's data and dialog-state concerns into two
dedicated hooks, and separate each *-dialog.tsx into a reusable form
component plus a thin dialog chrome wrapper.

- `useProductDetailData(id)` consolidates 5 queries + 8 mutations +
  derived `itineraryNameById` map + `invalidateProduct`.
- `useProductDetailDialogs()` replaces 6 scattered useState pairs with
  Toggle / EditingToggle<T> shapes that expose openNew / openEdit / close.
- Page drops 303 → 150 lines and becomes a pure presenter with no inline
  useQuery / useMutation / useState.
- 10 dialogs split into *-form.tsx (owns form logic + mutation + fields)
  and *-dialog.tsx (Sheet/Dialog chrome only): itinerary, day, option,
  option-price-rule, unit, unit-price-rule, service, detail, schedule,
  departure. Dialog files re-export data types so existing importers
  (options-section, options-pricing, detail-shared) keep working.
- Deletes dead `product-version-dialog.tsx` (zero importers).

Forms render inline action buttons rather than sticky SheetFooter bars,
matching the existing `product-form.tsx` pattern and keeping forms
dialog-agnostic so they can drop into wizards, drawers, or detail panels.

Closes #279 (points 2/3/4; point 1 was already shipped).
