+++ templates/operator/src/components/voyant/products/product-service-dialog.tsx
@@ -1,68 +1,9 @@
-import { useQuery } from "@tanstack/react-query"
-import { Loader2 } from "lucide-react"
-import { useEffect } from "react"
-import { useForm } from "react-hook-form"
-import { z } from "zod/v4"
-import {
-  Button,
-  Input,
-  Label,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-  Sheet,
-  SheetBody,
-  SheetContent,
-  SheetFooter,
-  SheetHeader,
-  SheetTitle,
-  Textarea,
-} from "@/components/ui"
+import { Sheet, SheetBody, SheetContent, SheetHeader, SheetTitle } from "@/components/ui"
 import { useAdminMessages } from "@/lib/admin-i18n"
-import { api } from "@/lib/api-client"
-import { zodResolver } from "@/lib/zod-resolver"
 
-type ServiceMessages = ReturnType<typeof useAdminMessages>["products"]["operations"]["services"]
+import { type DayServiceData, ServiceForm } from "./product-service-form"
 
-const buildServiceFormSchema = (messages: ServiceMessages) =>
-  z.object({
-    serviceType: z.enum(["accommodation", "transfer", "experience", "guide", "meal", "other"]),
-    name: z.string().min(1, messages.validationNameRequired),
-    description: z.string().optional().nullable(),
-    supplierServiceId: z.string().optional().nullable(),
-    costCurrency: z.string().min(3).max(3, messages.validationIsoCurrency),
-    costAmount: z.coerce.number().min(0, messages.validationCostNonNegative),
-    quantity: z.coerce.number().int().positive().default(1),
-    sortOrder: z.coerce.number().int().optional().or(z.literal("")).nullable(),
-    notes: z.string().optional().nullable(),
-  })
-
-type ServiceFormSchema = ReturnType<typeof buildServiceFormSchema>
-type ServiceFormValues = z.input<ServiceFormSchema>
-type ServiceFormOutput = z.output<ServiceFormSchema>
-
-type DayServiceData = {
-  id: string
-  serviceType: "accommodation" | "transfer" | "experience" | "guide" | "meal" | "other"
-  name: string
-  description: string | null
-  supplierServiceId: string | null
-  costCurrency: string
-  costAmountCents: number
-  quantity: number
-  sortOrder: number | null
-  notes: string | null
-}
-
-type SupplierServiceOption = {
-  id: string
-  supplierId: string
-  supplierName: string
-  serviceType: string
-  name: string
-}
+export type { DayServiceData }
 
 type ServiceDialogProps = {
   open: boolean
@@ -71,25 +12,6 @@ type ServiceDialogProps = {
   dayId: string
   service?: DayServiceData
   onSuccess: () => void
-}
-
-function getServiceTypeLabel(type: string, messages: ServiceMessages): string {
-  switch (type) {
-    case "accommodation":
-      return messages.serviceTypeAccommodation
-    case "transfer":
-      return messages.serviceTypeTransfer
-    case "experience":
-      return messages.serviceTypeExperience
-    case "guide":
-      return messages.serviceTypeGuide
-    case "meal":
-      return messages.serviceTypeMeal
-    case "other":
-      return messages.serviceTypeOther
-    default:
-      return type
-  }
 }
 
 export function ServiceDialog({
@@ -101,110 +23,8 @@ export function ServiceDialog({
   onSuccess,
 }: ServiceDialogProps) {
   const messages = useAdminMessages()
-  const productMessages = messages.products.core
   const serviceMessages = messages.products.operations.services
   const isEditing = !!service
-  const serviceFormSchema = buildServiceFormSchema(serviceMessages)
-  const serviceTypes = [
-    { value: "accommodation", label: serviceMessages.serviceTypeAccommodation },
-    { value: "transfer", label: serviceMessages.serviceTypeTransfer },
-    { value: "experience", label: serviceMessages.serviceTypeExperience },
-    { value: "guide", label: serviceMessages.serviceTypeGuide },
-    { value: "meal", label: serviceMessages.serviceTypeMeal },
-    { value: "other", label: serviceMessages.serviceTypeOther },
-  ] as const
-
-  // Fetch suppliers + their services for the picker
-  const { data: suppliersData } = useQuery({
-    queryKey: ["suppliers-for-picker"],
-    queryFn: async () => {
-      const res = await api.get<{
-        data: Array<{ id: string; name: string }>
-        total: number
-      }>("/v1/suppliers?limit=100")
-      const options: SupplierServiceOption[] = []
-      for (const supplier of res.data) {
-        const servicesRes = await api.get<{
-          data: Array<{ id: string; serviceType: string; name: string }>
-        }>(`/v1/suppliers/${supplier.id}/services`)
-        for (const svc of servicesRes.data) {
-          options.push({
-            id: svc.id,
-            supplierId: supplier.id,
-            supplierName: supplier.name,
-            serviceType: svc.serviceType,
-            name: svc.name,
-          })
-        }
-      }
-      return options
-    },
-    enabled: open,
-  })
-
-  const form = useForm<ServiceFormValues, unknown, ServiceFormOutput>({
-    resolver: zodResolver(serviceFormSchema),
-    defaultValues: {
-      serviceType: "accommodation",
-      name: "",
-      description: "",
-      supplierServiceId: "",
-      costCurrency: "EUR",
-      costAmount: 0,
-      quantity: 1,
-      sortOrder: "",
-      notes: "",
-    },
-  })
-
-  useEffect(() => {
-    if (open && service) {
-      form.reset({
-        serviceType: service.serviceType,
-        name: service.name,
-        description: service.description ?? "",
-        supplierServiceId: service.supplierServiceId ?? "",
-        costCurrency: service.costCurrency,
-        costAmount: service.costAmountCents / 100,
-        quantity: service.quantity,
-        sortOrder: service.sortOrder ?? "",
-        notes: service.notes ?? "",
-      })
-    } else if (open) {
-      form.reset()
-    }
-  }, [open, service, form])
-
-  const handleSupplierServiceSelect = (supplierServiceId: string | null) => {
-    const nextSupplierServiceId = supplierServiceId ?? ""
-    form.setValue("supplierServiceId", nextSupplierServiceId)
-    const option = suppliersData?.find((o) => o.id === nextSupplierServiceId)
-    if (option) {
-      form.setValue("name", option.name)
-      form.setValue("serviceType", option.serviceType as ServiceFormValues["serviceType"])
-    }
-  }
-
-  const onSubmit = async (values: ServiceFormOutput) => {
-    const payload = {
-      serviceType: values.serviceType,
-      name: values.name,
-      description: values.description || null,
-      supplierServiceId: values.supplierServiceId || null,
-      costCurrency: values.costCurrency,
-      costAmountCents: Math.round(values.costAmount * 100),
-      quantity: values.quantity,
-      sortOrder: values.sortOrder && typeof values.sortOrder === "number" ? values.sortOrder : null,
-      notes: values.notes || null,
-    }
-
-    if (isEditing) {
-      await api.patch(`/v1/products/${productId}/days/${dayId}/services/${service.id}`, payload)
-    } else {
-      await api.post(`/v1/products/${productId}/days/${dayId}/services`, payload)
-    }
-    onSuccess()
-  }
 
   return (
     <Sheet open={open} onOpenChange={onOpenChange}>
@@ -214,153 +34,15 @@ export function ServiceDialog({
             {isEditing ? serviceMessages.editTitle : serviceMessages.newTitle}
           </SheetTitle>
         </SheetHeader>
-        <form
-          onSubmit={form.handleSubmit(onSubmit)}
-          className="flex flex-1 flex-col overflow-hidden"
-        >
-          <SheetBody className="grid gap-4">
-            {/* Supplier service picker */}
-            {suppliersData && suppliersData.length > 0 && (
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.supplierServiceLabel}</Label>
-                <Select
-                  value={form.watch("supplierServiceId") ?? ""}
-                  onValueChange={handleSupplierServiceSelect}
-                  items={
-                    suppliersData?.map((opt) => ({
-                      value: opt.id,
-                      label: `${opt.supplierName} — ${opt.name} (${getServiceTypeLabel(
-                        opt.serviceType,
-                        serviceMessages,
-                      )})`,
-                    })) ?? []
-                  }
-                >
-                  <SelectTrigger className="w-full">
-                    <SelectValue placeholder={serviceMessages.supplierServicePlaceholder} />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {suppliersData.map((opt) => (
-                      <SelectItem key={opt.id} value={opt.id}>
-                        {opt.supplierName} — {opt.name} (
-                        {getServiceTypeLabel(opt.serviceType, serviceMessages)})
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-            )}
-
-            <div className="grid grid-cols-2 gap-4">
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.serviceTypeLabel}</Label>
-                <Select
-                  value={form.watch("serviceType")}
-                  onValueChange={(v) =>
-                    form.setValue("serviceType", v as ServiceFormValues["serviceType"])
-                  }
-                  items={serviceTypes}
-                >
-                  <SelectTrigger className="w-full">
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {serviceTypes.map((t) => (
-                      <SelectItem key={t.value} value={t.value}>
-                        {t.label}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.nameLabel}</Label>
-                <Input {...form.register("name")} placeholder={serviceMessages.namePlaceholder} />
-                {form.formState.errors.name && (
-                  <p className="text-xs text-destructive">{form.formState.errors.name.message}</p>
-                )}
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-2">
-              <Label>{serviceMessages.descriptionLabel}</Label>
-              <Textarea
-                {...form.register("description")}
-                placeholder={serviceMessages.descriptionPlaceholder}
-              />
-            </div>
-
-            <div className="grid grid-cols-3 gap-4">
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.costCurrencyLabel}</Label>
-                <Input
-                  {...form.register("costCurrency")}
-                  placeholder={serviceMessages.costCurrencyPlaceholder}
-                  maxLength={3}
-                  className="uppercase"
-                />
-                {form.formState.errors.costCurrency && (
-                  <p className="text-xs text-destructive">
-                    {form.formState.errors.costCurrency.message}
-                  </p>
-                )}
-              </div>
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.costAmountLabel}</Label>
-                <Input
-                  {...form.register("costAmount")}
-                  type="number"
-                  step="0.01"
-                  min="0"
-                  placeholder={serviceMessages.costAmountPlaceholder}
-                />
-                {form.formState.errors.costAmount && (
-                  <p className="text-xs text-destructive">
-                    {form.formState.errors.costAmount.message}
-                  </p>
-                )}
-              </div>
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.quantityLabel}</Label>
-                <Input
-                  {...form.register("quantity")}
-                  type="number"
-                  min="1"
-                  placeholder={serviceMessages.quantityPlaceholder}
-                />
-              </div>
-            </div>
-
-            <div className="grid grid-cols-2 gap-4">
-              <div className="flex flex-col gap-2">
-                <Label>{serviceMessages.sortOrderLabel}</Label>
-                <Input
-                  {...form.register("sortOrder")}
-                  type="number"
-                  placeholder={serviceMessages.sortOrderPlaceholder}
-                />
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-2">
-              <Label>{serviceMessages.notesLabel}</Label>
-              <Textarea
-                {...form.register("notes")}
-                placeholder={serviceMessages.notesPlaceholder}
-              />
-            </div>
-          </SheetBody>
-          <SheetFooter>
-            <Button type="button" variant="ghost" size="sm" onClick={() => onOpenChange(false)}>
-              {productMessages.cancel}
-            </Button>
-            <Button type="submit" size="sm" disabled={form.formState.isSubmitting}>
-              {form.formState.isSubmitting && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-              {isEditing ? productMessages.saveChanges : serviceMessages.create}
-            </Button>
-          </SheetFooter>
-        </form>
+        <SheetBody>
+          <ServiceForm
+            productId={productId}
+            dayId={dayId}
+            service={service}
+            onSuccess={onSuccess}
+            onCancel={() => onOpenChange(false)}
+          />
+        </SheetBody>
       </SheetContent>
     </Sheet>
   )
